Add opengraph/
Add an `opengraph/` directory that contains png images
optimized for the [opengraph protocol](https://ogp.me).

The opengraph protocol determines how a URL preview will
appear on Facebook, Mastodon and other popular websites.
Twitter supports the opengraph protocol to a certain extent,
but
[Twitter cards](https://developer.twitter.com/en/docs/twitter-for-websites/cards/guides/getting-started)
seem to be preferred and have better support.

The `bin/build-opengraph` shell script creates the contents
of the opengraph/ directory.